pyproject.toml
[tool.poetry]
name = "azure-logger"
version = "0.1.0"
description = "A simple logger to write to azure blob storage"
authors = ["Jesse Anderson <jesse.anderson@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "azure_logger"}]

[tool.poetry.dependencies]
python = "^3.8"
azure-storage-blob = ">=12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
